fix(tools): fold git log preview into shell
